fix: :bug: adjust physics engine
This commit introduces a SCALE_ADJUST. That value represents the number of pixels a unit is on the game. That was done so that we don't have to introduce misterious values when working with physics and, instead, we can focus on the math that works in the real world, because that will work by default.