feat: :sparkles: finish player movement
This commit introduces a very basic physics engine, where the skier (player) is pulled by gravity and is subject to friction (or air resistance).

When the direction of the skier changes, the previous velocity is projected onto the new direction.
This creates, accidentally, the intended behaviour of the player not sliding in the ice, even though there is no constraint explicitly enforcing that.

On a final note, the physics engine is most likely not finished and will be changed in the future. Nevertheless, since I am happy with its current state, I'm saying it is finished for now.