fix: :alembic: rotate velocity on player state change
On a previous commit, when a player's state changed, the new velocity vector would be the projection of the old one on the new direction.
I felt unhappy with this and I finally came to the conclusion that a better solution would be to simply rotate the velocity vector so that it would point in the new direction.

Therefore, in this commit, the new velocity is given as a rotation of the previous velocity vector.